requirements.txt
azure-core==1.33.0
azure-identity==1.21.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
isodate==0.7.2
msal==1.32.0
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
six==1.17.0
typing_extensions==4.13.1
urllib3==2.3.0
